+++ src/Gate.IO.Api/Wallet/Responses/GateWalletTransferId.cs
@@ -9,5 +9,5 @@ public record GateWalletTransferId
     /// Id
     /// </summary>
     [JsonProperty("id")]
-    public long Id { get; set; }
+    public string Id { get; set; }
 }